Tune default source IP address on all nodes - centralpoint, vrouters, standalones
On the nodes with the OpenVPN (all nodes with vrouter role) there is the additional internal VPN network. As the trafic is routed through the this network (local VPN gateway), default source address is selected also from the internal VPN network.

But we would rather used the source IP address from the LAN to not causing troubles for all the other nodes.